[moblab] Copy migration script to frontend.
The migration code make the assumption that the afe db and the
tko db are in separate database server.

The "migrate_info" table has no concept to differnetiate between
the two systems.

For the moment just use the AFE migrations for tko updates, the
lab it not using the sql database any more so this can be a safe
operation just for moblab.

TEST=copied the files on local moblab and ran migration.

Change-Id: Ibfda91f3dd241fd67ab7919b923debab67f4ed81
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/autotest/+/2166933
Reviewed-by: Keith Haddow <haddowk@chromium.org>
Commit-Queue: Keith Haddow <haddowk@chromium.org>
Tested-by: Keith Haddow <haddowk@chromium.org>