Stainless green: Move waiting for interrupt to server test.
We are hitting a problem where the update is finished before we can
start the interruption. This is likely happening because we wait for
the progress to be reached in the client test and then then need to
wait for control to come back to the server test. By ending the client
test early we should have lots of time to wait for the progress in the
server test.

Add some extra info to other failures to help debug.

BUG=chromium:1087347
TEST=autoupdate_Interruptions

Change-Id: If2f891aa95020fc3de7710575ce770bf1caec4b1
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/autotest/+/2219691
Tested-by: David Haddock <dhaddock@chromium.org>
Reviewed-by: Amin Hassani <ahassani@chromium.org>
Reviewed-by: Kyle Shimabukuro <kyleshima@chromium.org>
Commit-Queue: David Haddock <dhaddock@chromium.org>
(cherry picked from commit 5fbb073d8590534ab351f133c487d8d87e3cc95c)
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/autotest/+/2226547
Reviewed-by: David Haddock <dhaddock@chromium.org>