fingerprint: Re-enable biod and updater late in cleanup
The cleanup tries to restore the original firmware and initialize
entropy. If we re-enable biod, it can interfere with the initialization
of the entropy, so we delay re-enabling biod. The order of restoring
the updater technically doesn't matter but we move it together with
biod for simpler code.

BUG=none
TEST=suite:fingerprint passes except firmware_Fingerprint.RDP0

Change-Id: I04a6b4ac07a654b9c54412010cfed98ee96375ad
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/autotest/+/2477883
Tested-by: Yicheng Li <yichengli@chromium.org>
Reviewed-by: Tom Hughes <tomhughes@chromium.org>
Commit-Queue: Yicheng Li <yichengli@chromium.org>
(cherry picked from commit 947f3f34b23cc68b46f4e7f3622d2888039c0858)
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/autotest/+/2489032
Reviewed-by: Yicheng Li <yichengli@chromium.org>