[autotest]:  Fix issue in cts_P test_suite control.
The predicate logic required that the suite name in the control
file be exactly the same as the suite being run.

As we are moving to non letter cts suite in control files it now
has two suites cts and cts_P - this caused the predicate test
logic to fail and reject all control files.

Fix the logic.

BUG=chromium:1149840
TEST=tested by patching the control file in the devserver on moblab.

Change-Id: I923cf6538c6b442f07c72ad1cc64f61c3cd1d03c
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/autotest/+/2543478
Reviewed-by: Kazuhiro Inaba <kinaba@chromium.org>
Commit-Queue: Kazuhiro Inaba <kinaba@chromium.org>
Tested-by: Keith Haddow <haddowk@chromium.org>
Auto-Submit: Keith Haddow <haddowk@chromium.org>