[autotest][telemetry]: patch to resolve devserver correctly
Right now, the telemetry tests choose a devserver based on the build
name only. This causes an issue when the DUT is in Labs 2, 4, or 6 and
the devserver is chosen from a restricted subnet. This causes telemetry
test to fail. Supply hostname to TelemetrySetup to address this issue.
Also, the requests to the devservers are via HTTP directly. This is not
an issue if the chosen devserver is in Lab 2, 4, or 6 but may cause
problems if the devserver is in a restricted subnet. Add a 'backup'
method to run http requests on the devserver by first SSH-ing into it.
Backup because running such commands via subprocess is not ideal (and
hence should be avoided or given lower priority over direct HTTP
requests). Moreover, 99% of the telemetry tests are run on DUTs in
2, 4, 6 Labs (not by design so subject to change at any time), so
if we choose the devserver correctly, then executing HTTP commands
via requests should suffice.

BUG=chromium:1165407
TEST=[in chroot] autoserv -s --control-name telemetry_Benchmarks.rendering.desktop --host-info-subdir <path to host_info_store> -m <dut hostname> -l <board>/<version>/crosbolt_perf_perbuild/telemetry_Benchmarks.rendering.desktop --lab False --verify_job_repo_url -p --local-only-host-info True -r <results_dir> on DUTs in Lab 6 and Lab 15

Previous-Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/autotest/+/2660141
Commit-Queue: Sanika Kulkarni <sanikak@chromium.org>
Commit-Queue: Congbin Guo <guocb@chromium.org>
Tested-by: Sanika Kulkarni <sanikak@chromium.org>
Reviewed-by: Congbin Guo <guocb@chromium.org>
(cherry picked from commit 211a4ef65ed05dcebfd08dccb9386748c964b928)

Change-Id: Ic69cb60fa4953a5d59b3e446899870b5b1b55771
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/autotest/+/2719155
Tested-by: Congbin Guo <guocb@chromium.org>
Auto-Submit: Congbin Guo <guocb@chromium.org>
Reviewed-by: Sanika Kulkarni <sanikak@chromium.org>
Commit-Queue: Congbin Guo <guocb@chromium.org>