fingerprint_test: Flash FPMCU on cleanup only when necessary
FingerprintSigner test doesn't call setup_test() method, so
_dut_needs_reboot attribute isn't present. During cleanup,
_initialize_running_fw_version() is called. This function checks if
FPMCU firmware version is not correct or rollback block id is not 1
and calls flash_rw_ro_firmware() to re-flash FPMCU if necessary.
Because _dut_needs_reboot was not initialized, flash_rw_ro_firmware()
was not rebooting zork after flashing, so as a result test was reporting
that communication with FPMCU is not possible.

Another observation is that FingerprintSigner test doesn't modify
FPMCU flash really, so in fact there is nothing to clean up. The
problem was fixed by adding _need_fw_restore flag and setting it when
we suspect that the firmware needs to be restored or when we flashed
the firmware.

BUG=b:195952425
TEST=Run fingerprint test suite on zork (morphius or berknip)
     `test_that --board zork suite:fingerprint` and make sure that
     firmware_FingerprintSigner test passes.

Change-Id: I5a9862936befddbffd22aa1063a11876c7c8fb51
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/autotest/+/3220476
Tested-by: Patryk Duda <patrykd@google.com>
Reviewed-by: Tom Hughes <tomhughes@chromium.org>
Commit-Queue: Patryk Duda <patrykd@google.com>
(cherry picked from commit ff7b233091f810a27e7db3f37bc59b95a7f60a67)
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/autotest/+/3233104