firmware_Cr50DeepSleepStress: sleep longer and ignore probing errors
ti50 does not allow deep sleep until 60 seconds after the most recent
AP power on and it does not allow deep sleep until 10 seconds after the
most recent power off, which ever comes later. Due to this, we need to
sleep longer than 60 seconds when we are trying to ensure that ti50 is
going into deep sleep

Also, when firmware is probing GSC for DidVid, it is expected for both
ti50 and cr50 that the GSC does not ACK the address byte, which is
reported as an ERROR in AP firmware log. This is an expect case and
should not fail the test.

BUG=b:259134543,b:284983703
TEST=firmware_Cr50DeepSleepStress

Change-Id: Ic7baa64b4960dfeb05e3f9d1dea659ae2346a727
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/autotest/+/4093314
Commit-Queue: Edward Hill <ecgh@chromium.org>
Reviewed-by: Jett Rink <jettrink@chromium.org>
Reviewed-by: Mary Ruthven <mruthven@chromium.org>
Tested-by: Edward Hill <ecgh@chromium.org>
Reviewed-by: Derek Basehore <dbasehore@chromium.org>
(cherry picked from commit c5bf038fde301fc1720948561bf5ed00ca26da84)
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/autotest/+/4585110
Reviewed-by: Aadit Maheshkumar Modi <modia@chromium.org>
Commit-Queue: Mary Ruthven <mruthven@chromium.org>
Auto-Submit: Mary Ruthven <mruthven@chromium.org>
Tested-by: Mary Ruthven <mruthven@chromium.org>
Reviewed-by: Kevin Shelton <kmshelton@chromium.org>
Reviewed-by: Otabek Kasimov <otabek@google.com>