power_VideoCall: Fix regular expressions for CPU names
CL:4000042 ("power_VideoCall: Update test spec selection") did
a seemingly minor re-shuffling of the regular expressions that
are used to select the resolution and codec type for the video
camera based on the CPU.

In some lines in the middle of the regular expressions the
trailing '|' is missing, and the last lines have an excess '|',
which causes the expression matches any string. As a result the
wrong camera presets are selected for some boards/CPUs.

Add the missing '|'s and remove the excess ones.

BUG=b:310836058
TEST=test_that ${dedede_dut} power_VideoCall.FDO_25min
  manual testing with cpu names next to regular expressions

Change-Id: Ifd64c4a0c13a5f48bc4cf85a60240e6533b6f31f
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/autotest/+/5051862
Commit-Queue: Matthias Kaehlcke <mka@chromium.org>
Reviewed-by: Puthikorn Voravootivat <puthik@chromium.org>
Tested-by: Matthias Kaehlcke <mka@chromium.org>